video refact

+++ common/lib/logic/file_logic.dart
@@ -9,8 +9,12 @@ import 'package:common/repository/file_repository.dart';
 import 'package:common/repository/interface/file_interface.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_cache_manager/flutter_cache_manager.dart';
+import 'package:video_compress/video_compress.dart';
+import 'package:video_player/video_player.dart';
 
-final fileLogic = StateNotifierProvider.family<FileLogic, FileState, String>((ref, id) => FileLogic(ref.read, id));
+final fileLogic =
+    StateNotifierProvider.family.autoDispose<FileLogic, FileState, String>((ref, id) => FileLogic(ref.read, id));
 
 class FileLogic extends BaseLogic<FileState> {
   final String id;
@@ -20,19 +24,75 @@ class FileLogic extends BaseLogic<FileState> {
 
   User get user => read(authenticationLogic).currentUser;
 
+  VideoPlayerController? controller;
+
   @override
-  Future<void> initialize() async {}
+  Future<void> initialize({String url = ""}) async {
+    try {
+      final cachedFile = await DefaultCacheManager().getSingleFile(url);
+      state = state.copyWith(file: cachedFile);
+    } catch (e) {}
+  }
+
+  void videoInitalize() {
+    if (state.file == null) return;
+    controller = VideoPlayerController.file(state.file!)..initialize();
+    state = state.copyWith(controller: controller);
+    controller?.addListener(() {
+      if (controller?.value.position == controller?.value.duration) {
+        controller?.seekTo(const Duration(seconds: 0));
+      }
+      state = state.copyWith(
+        duration: controller?.value.duration ?? Duration.zero,
+        position: controller?.value.position ?? Duration.zero,
+        isPlaying: controller?.value.isPlaying ?? false,
+      );
+    });
+  }
+
+  @override
+  void dispose() {
+    controller?.removeListener(() {});
+    controller?.dispose();
+    state = state.copyWith(
+      controller: null,
+      isPlaying: false,
+      duration: Duration.zero,
+      position: Duration.zero,
+    );
+    controller = null;
+    super.dispose();
+  }
 
   Future<void> uploadFile(String name, File file, String type) async {
-    repo.uploadFile(user.id, name, file).listen((event) async {
-      state = state.copyWith(progress: (event.bytesTransferred / event.totalBytes));
-      if (event.state == TaskState.success) {
-        final url = await downloadFile(name, file);
-        sendFile(type, url);
-      } else if (event.state == TaskState.error) {
-        removeFile();
+    File? _file;
+    if (type == "video") {
+      try {
+        await VideoCompress.setLogLevel(0);
+        final mediaInfo = await VideoCompress.compressVideo(file.path, quality: VideoQuality.MediumQuality);
+        _file = mediaInfo?.file;
+      } catch (e) {
+        VideoCompress.cancelCompression();
       }
-    });
+    } else {
+      _file = file;
+    }
+    if (_file != null) {
+      repo.uploadFile(user.id, name, _file).listen((event) async {
+        state = state.copyWith(progress: (event.bytesTransferred / event.totalBytes));
+        if (event.state == TaskState.success) {
+          final url = await downloadFile(name, file);
+          final cachedFile = await DefaultCacheManager().getSingleFile(url);
+          if (type == "video") {
+            videoInitalize();
+          }
+          state = state.copyWith(file: cachedFile);
+          sendFile(type, url);
+        } else if (event.state == TaskState.error) {
+          removeFile();
+        }
+      });
+    }
   }
 
   Future<String> downloadFile(String name, File file) async {
@@ -48,7 +108,7 @@ class FileLogic extends BaseLogic<FileState> {
 
   Future<void> sendFile(String type, String url) async {
     await repo.sendFile(user.id, id, url, type);
-    read(chatLogic.notifier).sendPushNoti("${user.name} képet küldött önnek");
+    read(chatLogic.notifier).sendPushNoti("${user.name} ${type == 'image' ? 'képet' : 'videót'} küldött önnek");
   }
 
   Future<void> removeFile() async {
